Remove Translog interface
We only have one implementation of this interface which makes not much
sense. This commit removes the abstraction.
