+++ src/main/java/org/elasticsearch/index/translog/TranslogFile.java
@@ -17,27 +17,25 @@
  * under the License.
  */
 
-package org.elasticsearch.index.translog.fs;
+package org.elasticsearch.index.translog;
 
 import org.elasticsearch.common.bytes.BytesArray;
 import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.util.concurrent.ReleasableLock;
 import org.elasticsearch.index.shard.ShardId;
-import org.elasticsearch.index.translog.Translog;
-import org.elasticsearch.index.translog.TranslogException;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.concurrent.locks.ReadWriteLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 
-public abstract class FsTranslogFile extends FsChannelReader {
+public abstract class TranslogFile extends ChannelReader {
 
     protected final ShardId shardId;
     protected final ReleasableLock readLock;
     protected final ReleasableLock writeLock;
 
-    public FsTranslogFile(ShardId shardId, long id, ChannelReference channelReference) {
+    public TranslogFile(ShardId shardId, long id, ChannelReference channelReference) {
         super(id, channelReference);
         this.shardId = shardId;
         ReadWriteLock rwl = new ReentrantReadWriteLock();
@@ -46,22 +44,22 @@ public abstract class FsTranslogFile extends FsChannelReader {
     }
 
 
-    public static enum Type {
+    public enum Type {
 
         SIMPLE() {
             @Override
-            public FsTranslogFile create(ShardId shardId, long id, ChannelReference channelReference, int bufferSize) throws IOException {
-                return new SimpleFsTranslogFile(shardId, id, channelReference);
+            public TranslogFile create(ShardId shardId, long id, ChannelReference channelReference, int bufferSize) throws IOException {
+                return new SimpleTranslogFile(shardId, id, channelReference);
             }
         },
         BUFFERED() {
             @Override
-            public FsTranslogFile create(ShardId shardId, long id, ChannelReference channelReference, int bufferSize) throws IOException {
-                return new BufferingFsTranslogFile(shardId, id, channelReference, bufferSize);
+            public TranslogFile create(ShardId shardId, long id, ChannelReference channelReference, int bufferSize) throws IOException {
+                return new BufferingTranslogFile(shardId, id, channelReference, bufferSize);
             }
         };
 
-        public abstract FsTranslogFile create(ShardId shardId, long id, ChannelReference raf, int bufferSize) throws IOException;
+        public abstract TranslogFile create(ShardId shardId, long id, ChannelReference raf, int bufferSize) throws IOException;
 
         public static Type fromString(String type) {
             if (SIMPLE.name().equalsIgnoreCase(type)) {
@@ -78,7 +76,7 @@ public abstract class FsTranslogFile extends FsChannelReader {
     public abstract Translog.Location add(BytesReference data) throws IOException;
 
     /** reuse resources from another translog file, which is guaranteed not to be used anymore */
-    public abstract void reuse(FsTranslogFile other) throws TranslogException;
+    public abstract void reuse(TranslogFile other) throws TranslogException;
 
     /** change the size of the internal buffer if relevant */
     public abstract void updateBufferSize(int bufferSize) throws TranslogException;
@@ -90,19 +88,19 @@ public abstract class FsTranslogFile extends FsChannelReader {
     public abstract boolean syncNeeded();
 
     @Override
-    public FsChannelSnapshot newSnapshot() {
-        return new FsChannelSnapshot(immutableReader());
+    public ChannelSnapshot newSnapshot() {
+        return new ChannelSnapshot(immutableReader());
     }
 
     /**
      * returns a new reader that follows the current writes (most importantly allows making
      * repeated snapshots that includes new content)
      */
-    public FsChannelReader reader() {
+    public ChannelReader reader() {
         channelReference.incRef();
         boolean success = false;
         try {
-            FsChannelReader reader = new InnerReader(this.id, channelReference);
+            ChannelReader reader = new InnerReader(this.id, channelReference);
             success = true;
             return reader;
         } finally {
@@ -114,7 +112,7 @@ public abstract class FsTranslogFile extends FsChannelReader {
 
 
     /** returns a new immutable reader which only exposes the current written operation * */
-    abstract public FsChannelImmutableReader immutableReader();
+    abstract public ChannelImmutableReader immutableReader();
 
     boolean assertBytesAtLocation(Translog.Location location, BytesReference expectedBytes) throws IOException {
         ByteBuffer buffer = ByteBuffer.allocate(location.size);
@@ -126,7 +124,7 @@ public abstract class FsTranslogFile extends FsChannelReader {
      * this class is used when one wants a reference to this file which exposes all recently written operation.
      * as such it needs access to the internals of the current reader
      */
-    final class InnerReader extends FsChannelReader {
+    final class InnerReader extends ChannelReader {
 
         public InnerReader(long id, ChannelReference channelReference) {
             super(id, channelReference);
@@ -134,22 +132,22 @@ public abstract class FsTranslogFile extends FsChannelReader {
 
         @Override
         public long sizeInBytes() {
-            return FsTranslogFile.this.sizeInBytes();
+            return TranslogFile.this.sizeInBytes();
         }
 
         @Override
         public int totalOperations() {
-            return FsTranslogFile.this.totalOperations();
+            return TranslogFile.this.totalOperations();
         }
 
         @Override
         protected void readBytes(ByteBuffer buffer, long position) throws IOException {
-            FsTranslogFile.this.readBytes(buffer, position);
+            TranslogFile.this.readBytes(buffer, position);
         }
 
         @Override
-        public FsChannelSnapshot newSnapshot() {
-            return FsTranslogFile.this.newSnapshot();
+        public ChannelSnapshot newSnapshot() {
+            return TranslogFile.this.newSnapshot();
         }
     }
 }
