+++ src/test/java/org/elasticsearch/index/translog/AbstractTranslogTests.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.elasticsearch.index.translog.fs;
+package org.elasticsearch.index.translog;
 
 import org.apache.lucene.index.Term;
 import org.apache.lucene.util.IOUtils;
@@ -32,7 +32,6 @@ import org.elasticsearch.common.util.concurrent.ConcurrentCollections;
 import org.elasticsearch.index.Index;
 import org.elasticsearch.index.VersionType;
 import org.elasticsearch.index.shard.ShardId;
-import org.elasticsearch.index.translog.*;
 import org.elasticsearch.test.ElasticsearchTestCase;
 import org.hamcrest.Matchers;
 import org.junit.After;
@@ -64,7 +63,7 @@ public abstract class AbstractTranslogTests extends ElasticsearchTestCase {
 
     protected final ShardId shardId = new ShardId(new Index("index"), 1);
 
-    protected FsTranslog translog;
+    protected Translog translog;
     protected Path translogDir;
 
     @Override
@@ -103,7 +102,7 @@ public abstract class AbstractTranslogTests extends ElasticsearchTestCase {
     }
 
 
-    protected abstract FsTranslog create() throws IOException;
+    protected abstract Translog create() throws IOException;
 
 
     protected void addToTranslogAndList(Translog translog, ArrayList<Translog.Operation> list, Translog.Operation op) {
@@ -114,24 +113,24 @@ public abstract class AbstractTranslogTests extends ElasticsearchTestCase {
 
     public void testIdParsingFromFile() {
         long id = randomIntBetween(0, Integer.MAX_VALUE);
-        Path file = translogDir.resolve(FsTranslog.TRANSLOG_FILE_PREFIX + id);
-        assertThat(FsTranslog.parseIdFromFileName(file), equalTo(id));
+        Path file = translogDir.resolve(Translog.TRANSLOG_FILE_PREFIX + id);
+        assertThat(Translog.parseIdFromFileName(file), equalTo(id));
 
-        file = translogDir.resolve(FsTranslog.TRANSLOG_FILE_PREFIX + id + ".recovering");
-        assertThat(FsTranslog.parseIdFromFileName(file), equalTo(id));
+        file = translogDir.resolve(Translog.TRANSLOG_FILE_PREFIX + id + ".recovering");
+        assertThat(Translog.parseIdFromFileName(file), equalTo(id));
 
-        file = translogDir.resolve(FsTranslog.TRANSLOG_FILE_PREFIX + randomNonTranslogPatternString(1, 10) + id);
-        assertThat(FsTranslog.parseIdFromFileName(file), equalTo(-1l));
+        file = translogDir.resolve(Translog.TRANSLOG_FILE_PREFIX + randomNonTranslogPatternString(1, 10) + id);
+        assertThat(Translog.parseIdFromFileName(file), equalTo(-1l));
 
-        file = translogDir.resolve(randomNonTranslogPatternString(1, FsTranslog.TRANSLOG_FILE_PREFIX.length() - 1));
-        assertThat(FsTranslog.parseIdFromFileName(file), equalTo(-1l));
+        file = translogDir.resolve(randomNonTranslogPatternString(1, Translog.TRANSLOG_FILE_PREFIX.length() - 1));
+        assertThat(Translog.parseIdFromFileName(file), equalTo(-1l));
     }
 
     private static String randomNonTranslogPatternString(int min, int max) {
        String string;
         do {
             string = randomRealisticUnicodeOfCodepointLength(randomIntBetween(min, max));
-        } while (FsTranslog.PARSE_ID_PATTERN.matcher(string).matches());
+        } while (Translog.PARSE_ID_PATTERN.matcher(string).matches());
         return string;
     }
 
@@ -361,14 +360,14 @@ public abstract class AbstractTranslogTests extends ElasticsearchTestCase {
     }
 
 
-    public void assertFileIsPresent(FsTranslog translog, long id) {
+    public void assertFileIsPresent(Translog translog, long id) {
         if (Files.exists(translogDir.resolve(translog.getFilename(id)))) {
             return;
         }
         fail(translog.getFilename(id) + " is not present in any location: " + translog.location());
     }
 
-    public void assertFileDeleted(FsTranslog translog, long id) {
+    public void assertFileDeleted(Translog translog, long id) {
         assertFalse("translog [" + id + "] still exists", Files.exists(translog.location().resolve(translog.getFilename(id))));
     }
