+++ src/test/java/org/elasticsearch/index/translog/FsBufferedTranslogTests.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.elasticsearch.index.translog.fs;
+package org.elasticsearch.index.translog;
 
 import org.elasticsearch.common.settings.ImmutableSettings;
 import org.elasticsearch.common.util.BigArrays;
@@ -32,10 +32,10 @@ import java.io.IOException;
 public class FsBufferedTranslogTests extends AbstractTranslogTests {
 
     @Override
-    protected FsTranslog create() throws IOException {
-        return new FsTranslog(shardId,
+    protected Translog create() throws IOException {
+        return new Translog(shardId,
                 ImmutableSettings.settingsBuilder()
-                        .put("index.translog.fs.type", FsTranslogFile.Type.BUFFERED.name())
+                        .put("index.translog.fs.type", TranslogFile.Type.BUFFERED.name())
                         .put("index.translog.fs.buffer_size", 10 + randomInt(128 * 1024))
                         .build(),
                 BigArrays.NON_RECYCLING_INSTANCE, translogDir
